Move DownloadInfoRepository to data package
