+++ app/src/main/java/org/mozilla/rocket/chrome/di/ChromeModule.kt
@@ -18,8 +18,8 @@ import org.mozilla.rocket.chrome.data.MenuRepo
 import org.mozilla.rocket.chrome.domain.ReadNewMenuItemsUseCase
 import org.mozilla.rocket.chrome.domain.ShouldShowNewMenuItemHintUseCase
 import org.mozilla.rocket.download.DownloadIndicatorViewModel
-import org.mozilla.rocket.download.DownloadInfoRepository
 import org.mozilla.rocket.download.DownloadInfoViewModel
+import org.mozilla.rocket.download.data.DownloadInfoRepository
 import org.mozilla.rocket.helper.StorageHelper
 import org.mozilla.rocket.home.contenthub.domain.SetContentHubEnabledUseCase
 import org.mozilla.rocket.home.contenthub.domain.ShouldShowContentHubUseCase
