+++ app/src/main/java/org/mozilla/rocket/download/DownloadIndicatorViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import org.mozilla.rocket.download.data.DownloadInfoRepository
 
 class DownloadIndicatorViewModel(private val repository: DownloadInfoRepository) : ViewModel() {
